added biz logic to auth controller and reorganized file structure for BE
